+++ source/endpoints_types_impl.c
@@ -10,6 +10,24 @@
 #include <aws/sdkutils/private/endpoints_types_impl.h>
 #include <aws/sdkutils/private/endpoints_util.h>
 
+uint64_t aws_endpoints_fn_name_hash[AWS_ENDPOINTS_FN_LAST];
+
+void aws_endpoints_rule_engine_init() {
+    aws_endpoints_fn_name_hash[AWS_ENDPOINTS_FN_IS_SET] = aws_hash_c_string("isSet");
+    aws_endpoints_fn_name_hash[AWS_ENDPOINTS_FN_NOT] = aws_hash_c_string("not");
+    aws_endpoints_fn_name_hash[AWS_ENDPOINTS_FN_GET_ATTR] = aws_hash_c_string("getAttr");
+    aws_endpoints_fn_name_hash[AWS_ENDPOINTS_FN_SUBSTRING] = aws_hash_c_string("substring");
+    aws_endpoints_fn_name_hash[AWS_ENDPOINTS_FN_STRING_EQUALS] = aws_hash_c_string("stringEquals");
+    aws_endpoints_fn_name_hash[AWS_ENDPOINTS_FN_BOOLEAN_EQUALS] = aws_hash_c_string("booleanEquals");
+    aws_endpoints_fn_name_hash[AWS_ENDPOINTS_FN_URI_ENCODE] = aws_hash_c_string("uriEncode");
+    aws_endpoints_fn_name_hash[AWS_ENDPOINTS_FN_PARSE_URL] = aws_hash_c_string("parseURL");
+    aws_endpoints_fn_name_hash[AWS_ENDPOINTS_FN_IS_VALID_HOST_LABEL] = aws_hash_c_string("isValidHostLabel");
+    aws_endpoints_fn_name_hash[AWS_ENDPOINTS_FN_AWS_PARTITION] = aws_hash_c_string("aws.partition");
+    aws_endpoints_fn_name_hash[AWS_ENDPOINTS_FN_AWS_PARSE_ARN] = aws_hash_c_string("aws.parseArn");
+    aws_endpoints_fn_name_hash[AWS_ENDPOINTS_FN_AWS_IS_VIRTUAL_HOSTABLE_S3_BUCKET] =
+        aws_hash_c_string("aws.isVirtualHostableS3Bucket");
+}
+
 static void s_on_condition_array_element_clean_up(void *element) {
     struct aws_endpoints_condition *condition = element;
     aws_endpoints_condition_clean_up(condition);
@@ -149,4 +167,35 @@ void aws_endpoints_expr_clean_up(struct aws_endpoints_expr *expr) {
     }
 
     AWS_ZERO_STRUCT(*expr);
+}
+
+struct owning_cursor aws_endpoints_owning_cursor_create(struct aws_string *str) {
+    struct owning_cursor ret = {.string = str, .cur = aws_byte_cursor_from_string(str)};
+    return ret;
+}
+
+struct owning_cursor aws_endpoints_non_owning_cursor_create(struct aws_byte_cursor cur) {
+    struct owning_cursor ret = {.string = NULL, .cur = cur};
+    return ret;
+}
+
+static void s_eval_value_clean_up_cb(void *value);
+
+void aws_endpoints_eval_value_clean_up(struct eval_value *eval_value) {
+    if (eval_value->type == AWS_ENDPOINTS_EVAL_VALUE_STRING) {
+        aws_string_destroy(eval_value->v.string.string);
+    }
+
+    if (eval_value->type == AWS_ENDPOINTS_EVAL_VALUE_OBJECT) {
+        aws_string_destroy(eval_value->v.object.string);
+    }
+
+    if (eval_value->type == AWS_ENDPOINTS_EVAL_VALUE_ARRAY) {
+        aws_array_list_deep_clean_up(&eval_value->v.array, s_eval_value_clean_up_cb);
+    }
+}
+
+static void s_eval_value_clean_up_cb(void *value) {
+    struct eval_value *eval_value = value;
+    aws_endpoints_eval_value_clean_up(eval_value);
 }
